lab1_2/backend/lab1/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from mainapp.views import VideoGameViewSet

router = DefaultRouter()
router.register('games', VideoGameViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
]
