lab1_2/backend/requirements.txt
django
djangorestframework
psycopg[binary,pool]
python-dotenv
django-cors-headers
gunicorn
faker
